botofarm/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-multipart>=0.0.9
prometheus-fastapi-instrumentator>=6.1.0
